# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/validate_captcha": {
      "get": {
        "operationId": "UsersController_validateCaptcha",
        "parameters": [
          {
            "name": "api",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User Routes"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/users/generate_session": {
      "put": {
        "operationId": "UsersController_generateSession",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User Routes"
        ]
      }
    },
    "/users/vote": {
      "post": {
        "operationId": "UsersController_vote",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User Routes"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend Computação Amostra",
    "description": "API Routes and data",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token"
      }
    },
...
  }
}